+++ tests/conftest.py
@@ -1,7 +1,10 @@
 import sys
 
+import django
 import pytest
 from django.urls import reverse
+
+pytest.DJANGO41 = django.VERSION[0:2] == (4, 1)
 
 
 def pytest_configure(config):