+++ tests/test_admin.py
@@ -122,7 +122,11 @@ def test_archive_log(app, settings):
     LogEntryFactory(action_time=datetime.date(2000, 1, 1))
     res = app.get(url, user='sax')
     res = res.click("Archive")
-    res = res.form.submit()
+    if pytest.DJANGO41:
+        form = res.forms[1]
+    else:
+        form = res.form
+    res = form.submit()
     assert res.status_code == 302
     assert LogEntry.objects.count() == 1
 
@@ -133,8 +137,12 @@ def test_group_history(app, settings):
     g = GroupFactory()
     url = reverse(admin_urlname(Group._meta, 'change'), args=[g.id])
     res = app.get(url, user='sax')
-    res.form['permissions'] = [Permission.objects.first().pk]
-    res.form.submit()
+    if pytest.DJANGO41:
+        form = res.forms[1]
+    else:
+        form = res.form
+    form['permissions'] = [Permission.objects.first().pk]
+    res = form.submit()
 
     res = app.get(url, user='sax')
     res = res.click("History")
@@ -149,9 +157,13 @@ def test_user_history(app, settings):
     url = reverse(admin_urlname(User._meta, 'change'), args=[u.id])
 
     res = app.get(url, user='sax')
-    res.form['user_permissions'] = [Permission.objects.first().pk]
-    res.form['groups'] = [g.pk]
-    res.form.submit()
+    if pytest.DJANGO41:
+        form = res.forms[1]
+    else:
+        form = res.form
+    form['user_permissions'] = [Permission.objects.first().pk]
+    form['groups'] = [g.pk]
+    form.submit()
 
     res = app.get(url, user='sax')
     res = res.click("History")