+++ tests/test_console.py
@@ -21,7 +21,10 @@ def test_email_panel(app, mailoutbox):
     res = app.get(url, user='sax')
     res = res.click("Console")
     res = res.click("Test Email")
-    res = res.form.submit()
+    if pytest.DJANGO41:
+        res = res.forms[1].submit()
+    else:
+        res = res.form.submit()
     assert res.status_code == 200
     assert len(mailoutbox) == 1
     assert mailoutbox[0].subject == "Send email test: 'django.core.mail.send_mail'"